requirements.txt
kaggle>=1.7.0
pandas>=2.0.0
pillow>=10.0.0
requests>=2.25.0
tqdm>=4.60.0
